Consolidate AnVIL and HCA boilerplate (partial #5358)
